requirements.txt
aws-cdk.assets==1.134.0
aws-cdk.aws-acmpca==1.134.0
aws-cdk.aws-apigateway==1.134.0
aws-cdk.aws-applicationautoscaling==1.134.0
aws-cdk.aws-autoscaling-common==1.134.0
aws-cdk.aws-certificatemanager==1.134.0
aws-cdk.aws-cloudformation==1.134.0
aws-cdk.aws-cloudwatch==1.134.0
aws-cdk.aws-codeguruprofiler==1.134.0
aws-cdk.aws-codestarconnections==1.134.0
aws-cdk.aws-codestarnotifications==1.134.0
aws-cdk.aws-cognito==1.134.0
aws-cdk.aws-ec2==1.134.0
aws-cdk.aws-ecr==1.134.0
aws-cdk.aws-ecr-assets==1.134.0
aws-cdk.aws-efs==1.134.0
aws-cdk.aws-elasticloadbalancingv2==1.134.0
aws-cdk.aws-events==1.134.0
aws-cdk.aws-iam==1.134.0
aws-cdk.aws-kms==1.134.0
aws-cdk.aws-lambda==1.134.0
aws-cdk.aws-logs==1.134.0
aws-cdk.aws-route53==1.134.0
aws-cdk.aws-s3==1.134.0
aws-cdk.aws-s3-assets==1.134.0
aws-cdk.aws-sam==1.134.0
aws-cdk.aws-secretsmanager==1.134.0
aws-cdk.aws-signer==1.134.0
aws-cdk.aws-sns==1.134.0
aws-cdk.aws-sqs==1.134.0
aws-cdk.aws-ssm==1.134.0
aws-cdk.cloud-assembly-schema==1.134.0
aws-cdk.core==1.134.0
aws-cdk.custom-resources==1.134.0
aws-cdk.cx-api==1.134.0
aws-cdk.region-info==1.134.0
aws-cdk.aws-codebuild==1.134.0
aws-cdk.aws-codepipeline-actions==1.134.0
aws-cdk.aws-rds==1.134.0
aws-cdk.aws-ecs-patterns==1.134.0
aws-cdk.aws-ssm==1.134.0
cloudcomponents.cdk-github-webhook==1.46.0
cloudcomponents.cdk-secret-key==1.18.0
-e .
